Move static HotReloader Instantiate method from behaviour to the static class
